Remove vuex
