+++ src/components/VHeader/meta/VFilterButton.stories.mdx
@@ -6,7 +6,6 @@ import {
   Story,
 } from '@storybook/addon-docs'
 import VFilterButton from '~/components/VHeader/VFilterButton.vue'
-import Vuex from 'vuex'
 
 <Meta
   title="Components/VHeader/VFilterButton"
