+++ src/utils/search-query-transform.ts
@@ -13,11 +13,24 @@ import {
 import { ALL_MEDIA, SupportedSearchType } from '~/constants/media'
 
 import { getParameterByName } from '~/utils/url-params'
-import type {
-  ApiQueryFilters,
-  ApiQueryKeys,
-  ApiQueryParams,
-} from '~/store/types'
+
+export interface ApiQueryParams {
+  q?: string
+  license?: string
+  license_type?: string
+  extension?: string
+  size?: string
+  aspect_ratio?: string
+  searchBy?: string
+  categories?: string
+  source?: string
+  duration?: string
+  mature?: string
+  page?: string
+}
+
+export type ApiQueryFilters = Omit<ApiQueryParams, 'q'>
+export type ApiQueryKeys = keyof ApiQueryFilters
 
 const filterPropertyMappings: Record<FilterCategory, ApiQueryKeys> = {
   licenses: 'license',
