+++ test/unit/specs/components/related-audios.spec.js
@@ -1,5 +1,4 @@
 import { createLocalVue, mount } from '@vue/test-utils'
-import Vuex from 'vuex'
 import VueI18n from 'vue-i18n'
 
 import VRelatedAudio from '~/components/VAudioDetails/VRelatedAudio.vue'
@@ -9,7 +8,6 @@ import render from '../../test-utils/render'
 const audioResults = [{ id: 'audio1' }, { id: 'audio2' }]
 
 const localVue = createLocalVue()
-localVue.use(Vuex)
 localVue.use(VueI18n)
 localVue.prototype.$nuxt = {
   nbFetching: 0,
